+++ AlchemistCode/Cards/Token/Distillate.cs
@@ -12,7 +12,6 @@ public class Distillate : AlchemistCard
     public Distillate() : base(0, CardType.Skill, CardRarity.Token, TargetType.Self)
     {
         WithBlock(3, 2);
-        WithPower<RegenPower>(1, 1);
         WithCards(1, 0);
         WithKeyword(CardKeyword.Exhaust);
     }
@@ -20,7 +19,6 @@ public class Distillate : AlchemistCard
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {
         await CommonActions.CardBlock(this, play);
-        await CommonActions.ApplySelf<RegenPower>(choiceContext, this);
         await CommonActions.Draw(this, choiceContext);
     }
 }